fix(#139): move ageColor to lib/age-color-css.ts to fix react-refresh lint
