+++ src/components/pipeline/TaskCard.tsx
@@ -3,29 +3,7 @@ import { CSS } from '@dnd-kit/utilities';
 import type { Task, TransitionError } from '../../lib/types';
 import CopyButton from '../ui/CopyButton';
 import FlowStrip from './FlowStrip';
-
-export function ageColor(state: string, ageMinutes: number | null): string {
-  if (ageMinutes === null) return 'text-text-tertiary';
-
-  // State-specific thresholds [amber, red] in minutes
-  const thresholds: Record<string, [number, number]> = {
-    EXECUTION: [120, 480],
-    SETUP: [120, 480],
-    REVIEW_PENDING: [60, 180],
-    CI_PENDING: [60, 180],
-    AWAITING_OWNER: [60, 180],
-    CONTEXT: [240, 720],
-    RESEARCH: [240, 720],
-    DESIGN: [240, 720],
-    PLANNING: [240, 720],
-  };
-
-  const [amberThreshold, redThreshold] = thresholds[state] ?? [180, 480];
-
-  if (ageMinutes >= redThreshold) return 'text-red';
-  if (ageMinutes >= amberThreshold) return 'text-amber';
-  return 'text-text-tertiary';
-}
+import { ageColor } from '../../lib/age-color-css';
 
 function timeAgo(minutes: number | null): string {
   if (minutes === null || minutes === undefined) return '—';
